requirements.txt
requests>=2.31.0
openai
Anthropic
python-dotenv